backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain core packages
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5
langchain-huggingface==1.1.0

# Document loaders and text processing
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.11.8

# Supabase and vector store
supabase==2.10.0
vecs==0.4.0

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0

# OpenAI for embeddings and LLM
openai==1.12.0

# Optional: Anthropic Claude support
anthropic==0.18.1

# WebSockets for realtime connections
websockets>=13.0
